Replace select-frame-offset with find-next-frame-recursive
This is a checkin of the existing code I have on my desktop, to get
Git up to date before I check in a fix I made for frame-rowmajor.

In retrospect I don't know why I made this code change; it may have
been because I thought select-frame-offset wasn't workking right,
because sometimes the cursor would go to the wrong next frame.

This change removes the `arg` for goto-next-frame etc, so you can only
go to the adjacent frame.